In the previous checkin, we prefix each goose with the string "cytoscape" for the boss to get the name of the goose during workflow recording. This turns out to compromise the handleWorkflowAction function as Cytoscape cannot distinguish two network views with similar titles very well and sometimes won't create a new network when receiving a broadcast. We thus switch back to the previous naming mechanism. However, we append the application name string in the Broadcast calls so the recording feature still works correctly.
Another fix is we now support handling multiple data uris from the workflow.
This is to support the feature that user opens a file, and then broadcast some
gaggledata to that particular view. Details can be found in CyGoose.java

We also added logic to export the current network view and pass it to the Boss
to generate workflow session report. The logic is in HandleWorkflowAction.